requirements.txt
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-sock==0.7.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
priority==2.0.0
pycryptodomex==3.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.2
simple-websocket==1.0.0
sniffio==1.3.1
urllib3==2.2.1
watchdog==4.0.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yt-dlp==2024.5.27
